🚚 rename: Error 컴포넌트 이름을 ErrorPage로 변경 #46
