+++ src/pages/ErrorPage.tsx
@@ -1,3 +1,3 @@
-export default function Error() {
+export default function ErrorPage() {
   return <div>에러 페이지</div>;
 }
